BD-100: move identical classes in "util" to package matching structure in "dump", drop duplicate dependencies
