+++ dumpass/src/util/dumpass/GUI.java
@@ -66,8 +66,8 @@ import util.dump.ExternalizableBean.externalize;
 import util.dump.InfiniteGroupIndex;
 import util.dump.sort.InfiniteSorter;
 import util.dump.stream.SingleTypeObjectStreamProvider;
-import util.reflection.FieldAccessor;
-import util.reflection.FieldFieldAccessor;
+import util.dump.reflection.FieldAccessor;
+import util.dump.reflection.FieldFieldAccessor;
 import util.string.StringFilter;
 import util.swt.BorderPainter;
 import util.swt.CopyTableCellAction;
